B-1 refactoring path names.
